web_server/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
Flask-Login==0.6.2
Werkzeug==2.3.7
python-dotenv==1.0.0
email-validator==2.0.0
asyncio==3.4.3
aiofiles==23.2.1
python-multipart==0.0.6
celery==5.3.1
redis==5.0.1
blinker==1.6.2
pyjwt==2.8.0
bcrypt==4.0.1
pytest==7.4.2
pytest-flask==1.2.0
faker==19.12.0